Don't crash when the vigil window is made very small.
This is a workaround, not being able to get very small appearances is still
there. For now the main appearance won't shrink below a width of 10
or a height of 20, but thats okay since the interface is already impractical
at that point. The minimum appearance is then cropped to fit the window.